Command/speech_to_text/NeMo-speech-to-text/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
alabaster==0.7.12
apex==0.1
appdirs==1.4.3
ascii-graph==1.5.1
asn1crypto==1.2.0
async-timeout==3.0.1
attrdict==2.0.1
attrs==19.3.0
audioread==2.1.8
Babel==2.7.0
backcall==0.1.0
beautifulsoup4==4.8.1
black==19.10b0
bleach==3.1.0
bokeh==1.4.0
boto3==1.10.44
botocore==1.13.44
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
codecov==2.0.15
conda==4.8.0
conda-build==3.18.11
conda-package-handling==1.6.0
coverage==5.0
cryptography==2.8
ctc-decoders==1.1
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
DataProperty==0.43.1
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.9
docopt==0.6.2
docutils==0.15.2
easydict==1.9
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
frozendict==1.2
future==0.18.2
glob2==0.7
google-auth==1.10.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
h5py==2.10.0
html2text==2019.9.26
hypothesis==4.50.8
idna==2.8
idna-ssl==1.1.0
imageio==2.6.1
imagesize==1.1.0
importlib-metadata==1.3.0
inflect==3.0.2
ipdb==0.12.3
ipykernel==5.1.3
ipyparallel==6.2.4
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyter-server-proxy==1.3.2
jupyter-tensorboard==0.1.10
jupyterlab==1.0.4
jupyterlab-nvdashboard==0.2.1
jupyterlab-server==1.0.6
jupytext==1.3.0
kaldi-io==0.9.3
kiwisolver==1.1.0
latexcodec==2.0.0
libarchive-c==2.8
librosa==0.6.3
lief==0.9.0
llvmlite==0.28.0
lmdb==0.98
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.5.1
maskrcnn-benchmark==0.1
matplotlib==3.1.2
mbstrdecoder==0.8.1
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==3.0.5
more-itertools==8.0.2
msgfy==0.0.7
msgpack==0.6.1
msgpack-numpy==0.4.3.2
multidict==4.7.5
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
nemo-toolkit==0.10.0b9
networkx==2.0
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
num2words==0.5.10
numba==0.43.1
numpy==1.17.4
nvidia-dali==0.17.0
oauthlib==3.1.0
onnx==1.6.0
onnxruntime==1.2.0
opencv-python==3.4.1.15
oset==0.1.3
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parameterized==0.7.1
parso==0.5.2
pathspec==0.7.0
pathvalidate==0.29.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
Pillow-SIMD==5.3.0.post1
pip-api==0.0.14
pipreqs==0.4.10
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.4.3
pybtex==0.22.2
pybtex-docutils==0.2.2
pycocotools==2.0+nv0.3.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycuda==2019.1.2
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
pynvml==8.0.4
pyOpenSSL==19.1.0
pyparsing==2.4.5
pypinyin==0.37.0
pyrsistent==0.15.6
PySocks==1.7.1
pytablewriter==0.46.1
pytest==5.3.2
pytest-cov==2.8.1
pytest-pythonpath==0.7.3
pytest-runner==5.2
python-dateutil==2.8.1
python-hostlist==1.18
python-nvd3==0.15.0
python-slugify==4.0.0
pytools==2019.1.1
pytorch-transformers==1.1.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
regex==2018.1.10
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
revtok==0.0.3
rsa==4.0
ruamel-yaml==0.15.87
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.22
scipy==1.3.2
Send2Trash==1.5.0
sentencepiece==0.1.85
simpervisor==0.3
six==1.13.0
sklearn==0.0
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==1.9.5
sox==1.3.7
spacy==2.0.16
Sphinx==2.3.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-bibtex==1.0.0
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SSD==0.1
subword-nmt==0.3.3
tabledata==0.9.1
tabulate==0.8.6
tensorboard==2.1.0
tensorboardX==2.0
tensorrt==7.0.0.11
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
thinc==6.12.1
tokenizers==0.5.2
toml==0.10.0
toolz==0.10.0
torch==1.4.0a0+a5b4d78
torch-stft==0.1.4
torchtext==0.4.0
torchvision==0.5.0a0
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.3
transformers==2.6.0
typed-ast==1.4.1
typepy==0.6.0
typing==3.7.4.1
typing-extensions==3.7.4.1
ujson==1.35
Unidecode==1.1.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.10.11
yacs==0.1.6
yarg==0.1.9
yarl==1.4.2
youtokentome==1.0.6
zipp==0.6.0
